Update client.py
Changed how `price` is computed by using the formula: (bid_price+ask_price) / 2.
Changed the return value to the ratio of stock price_a to stock price_b.
For every iteration in the main method, we need to
store the datapoints we get from the `getDataPoint` method so that we can
properly call `getRatio` and print the right ratio.